requirements.txt
alembic==1.1.0
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.0
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.10.1
lib==3.0.0
Mako==1.1.0
MarkupSafe==1.1.1
mysqlclient==1.4.4
numpy==1.13.3
opencv-contrib-python-headless==3.4.4.19
python-dateutil==2.8.0
python-editor==1.0.4
PyYAML==5.1.2
pyzbar==0.1.8
RPi.GPIO==0.7.0
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.15.6
yamlordereddictloader==0.4.0
